VIC-3646 NEON optimized ApplyNoise (#2768)
Old: ~0.6ms
NEON: ~0.2ms

Also saves memory by storing the noise images as u8 instead of f32 which has the added benefit of reducing the number of neon instructions since I don't have to convert between u8 and float

+++ cannedAnimLib/proceduralFace/proceduralFaceDrawer.h
@@ -132,7 +132,7 @@ namespace Cozmo {
     static bool ConvertColorspace(const ProceduralFace& faceData, Vision::ImageRGB565& output, bool dirty);
 
 #if PROCEDURALFACE_NOISE_FEATURE
-    static const Array2d<f32>& GetNoiseImage(const Util::RandomGenerator& rng);
+    static const Array2d<u8>& GetNoiseImage(const Util::RandomGenerator& rng);
 #endif
 
 #if PROCEDURALFACE_SCANLINE_FEATURE
